pilab_setup.sh setup script for client OS Images
The script is to take a base image that is downloaded by the PiServer application and perform basic configuration (update, locale, time zone, etc.).  Generally it's best to create a clone of the downloaded image and work on the cloned image.  The script can be run from a CHROOTed terminal within the PiServer configuration tool.